# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API REST - Catégories, Produits et Acheteurs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/acheteurs": {
            "get": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Listes des acheteurs",
                "description": "Display a listing of the resource.",
                "operationId": "ceb448e7363a1b4f5c03c3887b9d3b33",
                "responses": {
                    "200": {
                        "description": "Liste des acheteurs"
                    }
                }
            },
            "post": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Créer un acheteur",
                "description": "Store a newly created resource in storage.",
                "operationId": "6b668a84a36dfc1380fc5739de88dcdc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "email"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Acheteur crée"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                }
            }
        },
        "/api/acheteurs/{id}": {
            "get": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Detail d'un acheteur",
                "description": "Display the specified resource.",
                "operationId": "0b4a046c39c051627363aa160b67ca12",
                "parameters": [
                    {
                        "name": "acheteur",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Acheteur trouvé"
                    },
                    "404": {
                        "description": "Acheteur non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Mettre à jour un acheteur",
                "description": "Update the specified resource in storage.",
                "operationId": "1b372799133848443208a82dd0578630",
                "parameters": [
                    {
                        "name": "Acheteur",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Acheteur mise à jour"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                }
            }
        },
        "/api/acheteurs/{acheteur}": {
            "delete": {
                "tags": [
                    "Acheteu\n    rs"
                ],
                "summary": "Supprimer un acheteur",
                "description": "Remove the specified resource from storage.",
                "operationId": "4bdb3b53d1f2bc34d5f2d236c7922237",
                "parameters": [
                    {
                        "name": "acheteur",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Acheteur supprimé"
                    }
                }
            }
        },
        "/api/acheteurs/{id}/acheter": {
            "post": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Enregistrer un achat",
                "operationId": "854d078e310f389e27dc8905decc67e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "produit_id",
                                    "quantite",
                                    "date_achat"
                                ],
                                "properties": {
                                    "produit_id": {
                                        "type": "integer"
                                    },
                                    "quantite": {
                                        "type": "integer"
                                    },
                                    "date_achat": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Achat enregistré"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                }
            }
        },
...
    },
...
}